+++ workspaces/app-defaults/plugins/app-react/src/drawer/index.ts
@@ -14,10 +14,14 @@
  * limitations under the License.
  */
 
-export { AppDrawerProvider, useAppDrawer } from './AppDrawerContext';
-export { ApplicationDrawer } from './ApplicationDrawer';
-export { DrawerPanel } from './DrawerPanel';
+export { useAppDrawer } from './hooks/useAppDrawer';
+export { ApplicationDrawer } from './components/ApplicationDrawer';
+export { DrawerPanel } from './components/DrawerPanel';
 
-export type { ApplicationDrawerProps } from './ApplicationDrawer';
-export type { DrawerPanelProps } from './DrawerPanel';
+export { appDrawerContentDataRef } from './extensions/appDrawerContentDataRef';
+export { AppDrawerContentBlueprint } from './extensions/AppDrawerContentBlueprint';
+export { appDrawerModule } from './extensions/appDrawerModule';
+
+export type { ApplicationDrawerProps } from './components/ApplicationDrawer';
+export type { DrawerPanelProps } from './components/DrawerPanel';
 export type { AppDrawerContent, AppDrawerApi } from './types';